+++ src/hazmat.rs
@@ -15,7 +15,7 @@ mod sieve;
 
 pub use lucas::{lucas_test, AStarBase, BruteForceBase, LucasBase, LucasCheck, SelfridgeBase};
 pub use miller_rabin::{minimum_mr_iterations, MillerRabin};
-pub use sieve::{random_odd_integer, SetBits, SieveFactory, SmallPrimesSieve, SmallPrimesSieveFactory};
+pub use sieve::{random_odd_integer, SetBits, SieveFactory, SmallFactorsSieve, SmallFactorsSieveFactory};
 
 use crypto_bigint::{Integer, Word};